src/requirements.txt
tensorflow==2.17.0
tensorflow-datasets
tensorflow-recommenders==0.7.3
google-cloud-aiplatform==1.60.0
google-cloud-storage==2.14.0
tensorboard
tensorboard-plugin-profile
tensorboard-plugin-wit
tensorboard-data-server
tensorflow-io